# [Performance] Mail key is now cached during player load