--- a/zone/zonedb.cpp
+++ b/zone/zonedb.cpp
@@ -1132,8 +1132,6 @@ bool ZoneDatabase::SaveCharacterData(
 		return false;
 	}
 
-	const auto mail_key = database.GetMailKey(c->CharacterID());
-
 	clock_t t = std::clock(); /* Function timer start */
 	const auto query = fmt::format(
 		"REPLACE INTO `character_data` ("
@@ -1427,7 +1425,7 @@ bool ZoneDatabase::SaveCharacterData(
 		m_epp->perAA,
 		m_epp->expended_aa,
 		m_epp->last_invsnapshot_time,
-		mail_key.c_str()
+		c->GetMailKeyFull()
 	);
 	auto results = database.QueryDatabase(query);
 	LogDebug(