There is quite a bit of idle query volume of players constantly querying for mail key when it barely changes. When players save, the mail key is queried, when UCS reconnects the mail key is queried from the DB.

This PR loads the mail key into memory on player load and re-uses the cache where referenced in the code

**Query Example**

```
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='687628' LIMIT 1 -- (1 row returned) (0.000318s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='736323' LIMIT 1 -- (1 row returned) (0.000143s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='732663' LIMIT 1 -- (1 row returned) (0.000369s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='734796' LIMIT 1 -- (1 row returned) (0.000119s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='736319' LIMIT 1 -- (1 row returned) (0.000104s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='736328' LIMIT 1 -- (1 row returned) (0.000127s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='737262' LIMIT 1 -- (1 row returned) (0.000258s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='737260' LIMIT 1 -- (1 row returned) (0.000152s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='736543' LIMIT 1 -- (1 row returned) (0.000203s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='735928' LIMIT 1 -- (1 row returned) (0.000320s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='721365' LIMIT 1 -- (1 row returned) (0.000148s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='736478' LIMIT 1 -- (1 row returned) (0.000098s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='736479' LIMIT 1 -- (1 row returned) (0.000110s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='731688' LIMIT 1 -- (1 row returned) (0.000113s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='730772' LIMIT 1 -- (1 row returned) (0.000299s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='736036' LIMIT 1 -- (1 row returned) (0.000097s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='735596' LIMIT 1 -- (1 row returned) (0.000097s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='733150' LIMIT 1 -- (1 row returned) (0.000226s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='736079' LIMIT 1 -- (1 row returned) (0.000356s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='736326' LIMIT 1 -- (1 row returned) (0.000406s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='736221' LIMIT 1 -- (1 row returned) (0.000335s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='728323' LIMIT 1 -- (1 row returned) (0.000108s)
5/7/2023 9:27:18 PM SELECT `mailkey` FROM `character_data` WHERE `id`='736325' LIMIT 1 -- (1 row returned) (0.000275s)
```

**Testing Validation**

![image](https://user-images.githubusercontent.com/3319450/236739278-14c1ef9b-e457-4a70-8f24-26669cb6714d.png)

Tested after the multiple keys refactor as well. Sent mail, chatted in UCS, saved, camped, reloaded, read E-Mail, sent E-Mail

![image](https://user-images.githubusercontent.com/3319450/236742131-700f17ba-08c6-433a-b7f8-77ca61941c6b.png)